api cleanup and changes for EOV grouping
