+++ ceda-web-api/utils/eovGrouping.json
@@ -1,0 +1,7 @@
+{
+  "carbon": ["inorganicCarbon", "surfaceCurrents"],
+  "currents": ["subSurfaceCurrents", "surfaceCurrents"],
+  "nutrients": ["nitrate", "phosphate", "sulphate", "silicate"],
+  "salinity": ["seaSurfaceSalinity", "subSurfaceSalinity"],
+  "temperature": ["seaSurfaceTemperature", "subSurfaceTemperature"]
+}
